+++ app/dev_mode/views.py
@@ -5,7 +5,7 @@ import time
 from flask import render_template, redirect, request, abort
 
 from app.authentication.user import UserConstants
-from app.jwt.jwt_encoder import Encoder
+from app.dev_mode.jwt_encoder import Encoder
 from app.schema_loader.schema_loader import available_schemas
 from . import dev_mode_blueprint
