+++ tests/app/jwt/test_jwt_decoder_using_rfc_example.py
@@ -1,7 +1,7 @@
 import unittest
 
 from app import settings
-from app.jwt.jwt_decoder import Decoder
+from app.authentication.jwt_decoder import Decoder
 
 # Not used in this test, but needed for the Decoder constructor
 private_pem = '''-----BEGIN RSA PRIVATE KEY-----
